Avoid execing for version detection in packaging
